as.matrix.bdsmatrix {bdsmatrix}	R Documentation

Convert a bdsmatrix to a ordinary (dense) matrix

Description

Method to convert from a Block Diagonal Sparse (bdsmatrix) matrix representation to an ordinary one

Usage

## S3 method for class 'bdsmatrix'
as.matrix(x, ...)

Arguments

x
a bdsmatrix object

...
other arguments are ignored (necessary to match the as.matrix template)

Details

Note that the conversion of a large bdsmatrix can easily exceed memory.

Value

a matrix

See Also

bdsmatrix

[Package bdsmatrix version 1.3-6 Index]