bdsI {bdsmatrix}	R Documentation

Sparse identity matrices

Description

This function will create an identitiy matrix, in the sparse bdsmatrix format.

Usage

bdsI(id, blocksize)

Arguments

id
the identifier list. This will become the dimnames of the final matrix, and must be a set of unique values. It's length determines the dimension of the final matrix

blocksize
the blocksize vector of the final matrix. If supplied, the sum of blocksizes must equal the dimension of the matrix. By default, the created matrix is as sparse as possible.

Value

an identity matrix.

Examples

imat <- bdsI(1:10)

[Package bdsmatrix version 1.3-6 Index]