bdsmatrix.reconcile {bdsmatrix}	R Documentation

Ensure alignment of two bdsmatrix objects

Description

This function is used by coxme. When a random effect is expressed as a sum of variance terms (matrices), it is important that all of them have the same row/column order and the same block structure. This does so, while retaining as much sparsity in the result as possible.

Usage

bdsmatrix.reconcile(varlist, group)

Arguments

varlist
a list, each element of which is a matrix or bdsmatrix object

group
a vector of dimnames, the target match for matrice's dimnames

Value

a varlist, whose individual elements may have had row/column rearrangment.

Author(s)

Terry Therneau

See Also

bdsmatrix

[Package bdsmatrix version 1.3-6 Index]