gchol-class {bdsmatrix}	R Documentation

Class "gchol"

Description

The result of a generalized Cholesky decomposition A=LDL' where A is a symmetric matrix, L is lower triangular with 1s on the diagonal, and D is a diagonal matrix.

Objects from the Class

These objects are created by the gchol function.

Slots

.Data:

A numeric vector containing the results of the decompostion

Dim:

An integer vector of length 2, the dimension of the matrix

Dimnames:

A list of length 2 containing the dimnames. These default to the dimnames of the matrix A

rank:

The rank of the matrix

Methods

%*%

signature(x = "gchol", y = "matrix"): multiply the cholesky decomposition by a matrix. That is, if A=LDL' is the decomposition, then gchol(A) %*% B will return L D^.5 B.

%*%

signature(x = "matrix", y = "gchol"): multiply by a matrix on the left

[

signature(x = "gchol"): if a square portion from the upper left corner is selected, then the result will be a gchol object, otherwise an ordinary matrix is returned. The latter most often occurs when printing part of the matrix at the command line.

coerce

signature(from = "gchol", to = "matrix"): Use of the as.matrix function will return L

diag

signature(x = "gchol"): Use of the diag function will return D

dim

signature(x = "gchol"): returns the dimension of the matrix

dimnames

signature(x = "gchol"): returns the dimnames

show

signature(object = "gchol"): By default a triangular matrix is printed showing D on the diagonal and L off the diagonal

gchol

signature(x= "matrix"): create a generalized Cholesky decompostion of the matrix

Note

The primary advantages of the genearlized decomposition, as compared to the standard chol function, has to do with redundant columns and generalized inverses (g-inverse). The lower triangular matrix L is always of full rank. The diagonal matrix D has a 0 element at position j if and only if the jth column of A is linearly dependent on columns 1 to j-1 preceding it. The g-inverse of A involves the inverse of L and a g-inverse of D. The g-inverse of D retains the zeros and inverts non-zero elements of D. This is very useful inside modeling functions such as coxph, since the X matrix can often contain a redundant column.

Author(s)

Terry Therneau

See Also

gchol

Examples

showClass("gchol")

[Package bdsmatrix version 1.3-6 Index]