gchol.bdsmatrix-class {bdsmatrix}	R Documentation

Class "gchol.bdsmatrix"

Description

Generalized cholesky decomposition of a bdsmatrix object, A= LDL' where A is symmetric, L is lower triangular with 1 on the diagonal, and D is diagonal.

Objects from the Class

These are created by the gchol function.

Slots

blocksize:

Integer vector of block sizes

blocks:

Numeric vector containing the blocks

rmat:

Dense portion of the decomposition

rank:

The rank of A

Dim:

Integer vector of length 2 containing the dimension

Dimnames:

List of length 2 containing the dimnames

Methods

%*%

signature(x = "gchol.bdsmatrix", y = "matrix"): ...

%*%

signature(x = "gchol.bdsmatrix", y = "numeric"): ...

%*%

signature(x = "matrix", y = "gchol.bdsmatrix"): ...

%*%

signature(x = "numeric", y = "gchol.bdsmatrix"): ...

[

signature(x = "gchol.bdsmatrix"): ...

coerce

signature(from = "gchol.bdsmatrix", to = "matrix"): ...

diag

signature(x = "gchol.bdsmatrix"): ...

dim

signature(x = "gchol.bdsmatrix"): ...

show

signature(object = "gchol.bdsmatrix"): ...

Note

The Cholesky decompostion of a block diagonal symmetric matrix is also block diagonal symmetric, so is stored in the same manner as a bdsmatrix object

Author(s)

Terry Therneau

See Also

bdsmatrix, gchol

Examples

showClass("gchol.bdsmatrix")

[Package bdsmatrix version 1.3-6 Index]